requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
h11==0.14.0
idna==3.10
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
redis==5.2.0
rich==13.9.4
richapi==0.3.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.0
typing_extensions==4.12.2
uvicorn==0.32.1
